feat: Implement a Drawer
