+++ app/src/main/java/com/nilhcem/blenamebadge/ui/fragments/main_text/MainTextDrawableNavigator.kt
@@ -1,7 +1,10 @@
 package com.nilhcem.blenamebadge.ui.fragments.main_text
 
+import android.widget.LinearLayout
+import android.widget.TextView
 import com.nilhcem.blenamebadge.data.DrawableInfo
 import com.nilhcem.blenamebadge.data.device.model.DataToSend
+import pl.droidsonroids.gif.GifImageView
 
 interface MainTextDrawableNavigator {
     fun selectText()
@@ -25,4 +28,8 @@ interface MainTextDrawableNavigator {
     fun showFileOverrideDialog(fileName: String, jsonString: String)
 
     fun saveFile(fileName: String, jsonString: String)
+    fun setupSpeedKnob()
+    fun configureEffects()
+    fun setBackgroundOf(card: LinearLayout?, image: GifImageView?, title: TextView?, checked: Boolean)
+    fun setupTabLayout()
 }
