--- app/src/main/res/color/navigation_item.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="@color/colorPrimary" android:state_selected="true" />
-    <item android:color="@color/lighter_grey" android:state_selected="false" />
-</selector>
