--- app/src/main/res/drawable/ic_add.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="48"
-    android:viewportHeight="48">
-    <path
-        android:fillColor="#ffffff"
-        android:pathData="M38,26H26v12h-4V26H10v-4h12V10h4v12h12v4z" />
-</vector>
