--- app/src/main/res/drawable/ic_bottom_bar_saved.xml
@@ -1,6 +1,0 @@
-<vector android:height="24dp" android:viewportHeight="32"
-    android:viewportWidth="32" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M7.5,10.5h17c1.1,0 2,-0.9 2,-2s-0.9,-2 -2,-2h-17c-1.1,0 -2,0.9 -2,2S6.4,10.5 7.5,10.5z"/>
-    <path android:fillColor="#FF000000" android:pathData="M24.5,13.5h-17c-1.1,0 -2,0.9 -2,2c0,1.1 0.9,2 2,2h17c1.1,0 2,-0.9 2,-2C26.5,14.4 25.6,13.5 24.5,13.5z"/>
-    <path android:fillColor="#FF000000" android:pathData="M24.5,20.5h-17c-1.1,0 -2,0.9 -2,2s0.9,2 2,2h17c1.1,0 2,-0.9 2,-2S25.6,20.5 24.5,20.5z"/>
-</vector>
