--- app/src/main/res/drawable/ic_bottom_bar_text.xml
@@ -1,8 +1,0 @@
-<vector android:height="24dp" android:viewportHeight="100"
-    android:viewportWidth="100" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M32.4,47.1h5.8v5.8h-5.8z"/>
-    <path android:fillColor="#FF000000" android:pathData="M61.8,47.1h5.8v5.8h-5.8z"/>
-    <path android:fillColor="#FF000000" android:pathData="M12.8,72.5h9.8v9.8h5.8v-9.8h43.2v9.8h5.8v-9.8h9.8V52.9H97V27.5h-5.8v19.6H77.4v-9.8h-9.8v-9.8h-5.8v9.8H38.2v-9.8h-5.8v9.8h-9.8v9.8H8.8V27.5H3.1v25.4h9.8L12.8,72.5L12.8,72.5zM57.8,43.1h13.8v13.8H57.8V43.1zM28.4,43.1h13.8v13.8H28.4V43.1z"/>
-    <path android:fillColor="#FF000000" android:pathData="M22.6,17.7h5.8v5.8h-5.8z"/>
-    <path android:fillColor="#FF000000" android:pathData="M71.6,17.7h5.8v5.8h-5.8z"/>
-</vector>
