+++ app/src/main/res/drawable/ic_menu_about.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="846.6"
+    android:viewportWidth="846.7" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M333.6,211.8c-18.1,0.9 -23.6,5.2 -25.3,23.8c-5,56.6 -7.9,135.3 20.7,188c-16.4,2.1 -32.6,4.7 -48.4,7.8c15.2,-41.1 13,-94.8 9.5,-134.8c-1.3,-14.9 -5.7,-18.4 -20.2,-19c-19.1,17 -44.3,27.2 -71.9,27.2c-27.6,0 -52.7,-10.3 -71.8,-27.2c-14.5,0.7 -18.9,4.1 -20.2,19c-4.6,52.6 -7.1,129.1 29.7,169.2c3.1,3.4 5.8,8 8.2,13.6C19.3,550.8 28.8,666.1 172.5,730.7c139.8,62.8 362,62.8 501.7,0c143.7,-64.6 153.2,-180 28.6,-251.3c2.3,-5.6 5,-10.2 8.2,-13.6c36.8,-40.1 34.3,-116.6 29.7,-169.2c-1.3,-14.9 -5.7,-18.4 -20.2,-19c-19.1,17 -44.3,27.2 -71.8,27.2s-52.7,-10.3 -71.8,-27.2c-14.5,0.7 -18.9,4.1 -20.2,19c-3.5,40 -5.8,93.7 9.5,134.8c-15.8,-3.1 -31.9,-5.8 -48.3,-7.8c28.6,-52.7 25.7,-131.4 20.7,-188c-1.6,-18.6 -7.1,-22.9 -25.2,-23.8c-23.9,21.2 -55.3,34 -89.7,34S357.5,232.9 333.6,211.8L333.6,211.8zM156.5,536.1c4.5,35 6.6,73 12.2,92.9c3.2,11.6 55.6,11.6 58.9,0c7.8,-28.1 8.8,-91.8 19.3,-132.6c34.9,-10.2 73.4,-17 113.1,-20.5c15.5,50.1 16.3,137.6 26.6,174.8c4,14.5 69.5,14.5 73.5,0c10.3,-37.2 11.2,-124.7 26.6,-174.8c39.7,3.5 78.2,10.4 113.1,20.5c10.4,40.8 11.5,104.6 19.3,132.6c3.2,11.6 55.6,11.6 58.9,0c5.5,-20 7.7,-57.9 12.2,-92.9c69.4,46.2 56.6,101.6 -38.5,144.3c-125.1,56.2 -331.7,56.2 -456.8,0C99.9,637.7 87.1,582.2 156.5,536.1zM198.1,163.1c32.6,0 59,26.4 59,59s-26.4,59 -59,59c-32.6,0 -59,-26.4 -59,-59S165.5,163.1 198.1,163.1L198.1,163.1zM648.6,163.1c32.6,0 59,26.4 59,59s-26.4,59 -59,59s-59,-26.4 -59,-59S616,163.1 648.6,163.1L648.6,163.1zM423.3,68.8c40.7,0 73.7,33 73.7,73.7s-33,73.7 -73.7,73.7c-40.7,0 -73.7,-33 -73.7,-73.7S382.6,68.8 423.3,68.8z"/>
+</vector>
