+++ app/src/main/res/drawable/ic_menu_bug.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="16"
+    android:viewportWidth="16" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M4.3,0.3c-0.3,0 -0.5,0.3 -0.4,0.6s0.3,0.5 0.6,0.4c0.4,0 0.6,0.1 0.9,0.4s0.5,0.6 0.7,1C6.1,2.9 6.1,3 6.2,3.2C5.5,4 5.2,5.1 5.8,6c0,0 0,0 0,0L3.3,4.4C3.2,4.3 3,4.3 2.9,4.3C2.6,4.3 2.5,4.6 2.5,4.9c0,0.1 0.1,0.2 0.2,0.3L5,6.7C4.8,7 4.6,7.3 4.4,7.6L0.8,9.4C0.5,9.4 0.4,9.7 0.4,10c0.1,0.3 0.3,0.4 0.6,0.4c0.1,0 0.1,0 0.2,-0.1l2.6,-1.3c-0.2,0.6 -0.3,1.2 -0.3,1.8c0,0 0,0.1 0,0.1l-1.9,2.6c-0.2,0.2 -0.1,0.5 0.1,0.7c0.2,0.2 0.5,0.1 0.7,-0.1c0,0 0,0 0,0l1.3,-1.7c0.4,1.6 1.4,2.9 2.7,3.5c0.6,-0.4 1.1,-1.2 1.1,-2.1v-6c0,-0.3 0.2,-0.5 0.5,-0.5s0.5,0.2 0.5,0.5v6c0,0.9 0.5,1.7 1.1,2.1c1.3,-0.6 2.3,-1.9 2.7,-3.5l1.3,1.7c0.2,0.2 0.5,0.3 0.7,0.1c0.2,-0.2 0.3,-0.5 0.1,-0.7c0,0 0,0 0,0l-1.9,-2.6c0,0 0,-0.1 0,-0.1c0,-0.6 -0.1,-1.3 -0.3,-1.8l2.6,1.3c0.2,0.2 0.5,0.1 0.7,-0.1c0.2,-0.2 0.1,-0.5 -0.1,-0.7c-0.1,0 -0.1,-0.1 -0.2,-0.1l-3.6,-1.8C11.4,7.3 11.2,7 11,6.7l2.3,-1.5c0.2,-0.2 0.3,-0.5 0.1,-0.7c-0.1,-0.2 -0.3,-0.3 -0.5,-0.2c-0.1,0 -0.1,0 -0.2,0.1L10.3,6c0,0 0,0 0,0c0.7,-1 0.2,-2.4 -0.8,-3.2c0,0 0,-0.1 0,-0.1c0.2,-0.4 0.4,-0.8 0.7,-1s0.5,-0.4 0.9,-0.4c0.3,0 0.5,-0.2 0.5,-0.5S11.3,0.3 11,0.3c-0.6,0 -1.2,0.3 -1.6,0.7C9,1.4 8.7,1.9 8.5,2.4C8.4,2.3 8.2,2.3 8,2.3c-0.3,0 -0.7,0.1 -1,0.2C7,2.5 7,2.4 7,2.4C6.8,1.9 6.5,1.4 6.1,1S5.1,0.3 4.5,0.3C4.5,0.3 4.4,0.3 4.3,0.3C4.4,0.3 4.4,0.3 4.3,0.3L4.3,0.3z"/>
+</vector>
