+++ app/src/main/res/drawable/ic_menu_buy.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="100"
+    android:viewportWidth="100" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M80,5H55L5,55l40,40l50,-50V20L80,5zM52.1,68.9L52.1,75l-4,0l0.1,-5.8c-4.1,0 -8.6,-1.1 -11.6,-2.6l1.9,-3.5c2,1.1 5.9,2.2 9.9,2.2c6.4,0 10.4,-2.5 10.4,-6.5c0,-2.8 -1.4,-5 -8.5,-7.2c-5.8,-1.8 -11.7,-4.4 -11.7,-10.7c0,-4.1 3.1,-8.4 9.9,-9.5l0.1,-6.2l4,0l-0.1,5.9c4.7,0.1 8,1.2 9.5,2l-1.7,3.6c-1,-0.5 -3.9,-1.6 -8.3,-1.6c-5.5,0 -9.4,2.4 -9.4,5.8c0,2.5 1.5,4.6 8.9,6.9c5.6,1.8 11.3,4.5 11.3,11C62.8,62.9 59.9,67.7 52.1,68.9zM80,25c-2.8,0 -5,-2.2 -5,-5s2.2,-5 5,-5s5,2.2 5,5S82.8,25 80,25z"/>
+</vector>
