+++ app/src/main/res/drawable/ic_menu_create.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:viewportHeight="64"
+    android:viewportWidth="64" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M12.9,46.4l-1.8,-1.8L9,53.7c-0.1,0.3 0,0.7 0.3,1c0.3,0.3 0.6,0.4 1,0.3l9.1,-2.1l-1.8,-1.8L12.9,46.4z"/>
+    <path android:fillColor="#FF000000" android:pathData="M54.7,16.7l-7.4,-7.4c-0.4,-0.4 -1.1,-0.4 -1.5,0L12.3,42.8l1.5,1.5l6,6l1.5,1.5l33.5,-33.5C55.1,17.8 55.1,17.2 54.7,16.7z"/>
+</vector>
