+++ app/src/main/res/drawable/ic_menu_save.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="100"
+    android:viewportHeight="100">
+    <group
+        android:pivotX="50"
+        android:pivotY="50"
+        android:scaleX="0.9"
+        android:scaleY="0.9">
+        <path
+            android:fillColor="#FF000000"
+            android:pathData="M43.2,5h11.5c0.5,0 1,0.4 1,1v24.3c0,0.5 -0.4,1 -1,1l0,0H43.2c-0.5,0 -1,-0.4 -1,-1c0,0 0,0 0,0V6C42.2,5.4 42.7,5 43.2,5C43.2,5 43.2,5 43.2,5z" />
+        <path
+            android:fillColor="#FF000000"
+            android:pathData="M93.5,25L75,6.5C74,5.5 72.8,5 71.5,5h-7v27.4c0,1.9 -1.5,3.4 -3.4,3.4c0,0 0,0 0,0H24.4c-1.9,0 -3.4,-1.5 -3.4,-3.4c0,0 0,0 0,0V5H10.5C7.5,5 5,7.5 5,10.5c0,0 0,0 0,0v79c0,3 2.5,5.5 5.5,5.5h79c3,0 5.5,-2.5 5.5,-5.5c0,0 0,0 0,0c0,0 0,-50.1 0,-61C95,27.2 94.5,26 93.5,25zM79,83c0,1.9 -1.5,3.4 -3.4,3.4c0,0 0,0 0,0H24.4c-1.9,0 -3.4,-1.5 -3.4,-3.4c0,0 0,0 0,0V46.9c0,-1.9 1.5,-3.4 3.4,-3.4c0,0 0,0 0,0h51.3c1.9,0 3.4,1.5 3.4,3.4c0,0 0,0 0,0V83z" />
+        <path
+            android:fillColor="#FF000000"
+            android:pathData="M31.7,55.9h36.7c0.4,0 0.6,0.3 0.6,0.6v3.7c0,0.4 -0.3,0.6 -0.6,0.6H31.7c-0.4,0 -0.6,-0.3 -0.6,-0.6v-3.7C31,56.2 31.3,55.9 31.7,55.9z" />
+        <path
+            android:fillColor="#FF000000"
+            android:pathData="M31.7,68.9h36.7c0.4,0 0.6,0.3 0.6,0.6v0v3.7c0,0.4 -0.3,0.6 -0.6,0.6h0H31.7c-0.4,0 -0.6,-0.3 -0.6,-0.6v-3.7C31,69.2 31.3,68.9 31.7,68.9L31.7,68.9L31.7,68.9z" />
+    </group>
+</vector>
