--- app/src/main/res/layout-land/activity_main.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.main.MainActivity">
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginBottom="@dimen/scroll_bar_bottom_margin"
-        android:fillViewport="true">
-
-        <LinearLayout
-            android:id="@+id/mainLayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <com.nilhcem.blenamebadge.ui.custom.PreviewBadge
-                android:id="@+id/preview_badge"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
-
-            <com.nilhcem.blenamebadge.ui.custom.CustomViewPager
-                android:id="@+id/viewPager"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="1">
-
-            </com.nilhcem.blenamebadge.ui.custom.CustomViewPager>
-        </LinearLayout>
-    </ScrollView>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:background="@android:color/white"
-        android:foreground="?attr/selectableItemBackground"
-        android:visibility="invisible"
-        app:itemHorizontalTranslationEnabled="false"
-        app:itemIconTint="@color/navigation_item"
-        app:itemTextColor="@color/navigation_item"
-        app:labelVisibilityMode="selected"
-        app:menu="@menu/navigation" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
