--- app/src/main/res/layout/activity_main.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.main.MainActivity">
-
-    <LinearLayout
-        android:id="@+id/mainLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <com.nilhcem.blenamebadge.ui.custom.PreviewBadge
-            android:id="@+id/preview_badge"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-
-        <androidx.viewpager.widget.ViewPager
-            android:id="@+id/viewPager"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1">
-
-        </androidx.viewpager.widget.ViewPager>
-    </LinearLayout>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="invisible"
-        android:layout_gravity="bottom"
-        android:background="@android:color/white"
-        android:foreground="?attr/selectableItemBackground"
-        app:itemHorizontalTranslationEnabled="false"
-        app:itemIconTint="@color/navigation_item"
-        app:itemTextColor="@color/navigation_item"
-        app:labelVisibilityMode="selected"
-        app:menu="@menu/navigation" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
