--- app/src/main/res/layout/spinner_item.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TextView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="@dimen/spacing_normal"
-    android:textColor="@android:color/black"
-    android:textSize="@dimen/text_size_large"/>
