--- app/src/main/res/menu/navigation.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/textDrawFragment"
-        android:icon="@drawable/ic_bottom_bar_text"
-        android:title="@string/text_draw_fragment" />
-
-    <item
-        android:id="@+id/savedFragment"
-        android:icon="@drawable/ic_bottom_bar_saved"
-        android:title="@string/saved_fragment" />
-
-</menu>
