+++ app/src/main/res/menu/save_menu.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/menu_import"
-        android:icon="@drawable/ic_add"
-        android:title="@string/menu_import"
-        app:showAsAction="always" />
+    <group android:id="@+id/saved_group">
+        <item
+            android:id="@+id/menu_import"
+            android:title="@string/menu_import"
+            app:showAsAction="always" />
+    </group>
 </menu>
