+++ app/src/main/res/values/dimens.xml
@@ -5,7 +5,6 @@
     <dimen name="card_elevation">2dp</dimen>
 
     <dimen name="text_size_normal">16sp</dimen>
-    <dimen name="text_size_large">18sp</dimen>
     <dimen name="text_size_large_extended">20sp</dimen>
 
     <dimen name="spacing_small">4dp</dimen>
@@ -13,7 +12,6 @@
     <dimen name="spacing_large">16dp</dimen>
 
     <dimen name="button_padding_left_right">42dp</dimen>
-    <dimen name="fab_margin">16dp</dimen>
     <dimen name="bottom_bar_margin">32dp</dimen>
     <dimen name="scroll_bar_bottom_margin">64dp</dimen>
     <dimen name="drawable_image_margin">5dp</dimen>
@@ -21,4 +19,10 @@
     <dimen name="title_min_height_large">40dp</dimen>
     <dimen name="knob_margin">20dp</dimen>
     <dimen name="knob_width">150dp</dimen>
+
+    <!-- Default screen margins, per the Android Design guidelines. -->
+    <dimen name="activity_horizontal_margin">16dp</dimen>
+    <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="nav_header_vertical_spacing">8dp</dimen>
+    <dimen name="nav_header_height">176dp</dimen>
 </resources>
